Séparer les modèles pydantic de l'api
