--- .devcontainer/devcontainer.json
@@ -1,19 +1,0 @@
-{
-	"name": "MSPR ETL - Dev Container",
-	"dockerComposeFile": "../docker-compose.yml",
-	"service": "frontend-dev",
-	"workspaceFolder": "/app",
-	"runServices": ["db", "grafana", "backend"],
-	"forwardPorts": [8000, 5432, 3000, 5173],
-	"postCreateCommand": "pip install --no-cache-dir -r requirements.txt && apk add git",
-	"customizations": {
-		"vscode": {
-			"extensions": [
-				"ms-python.python",
-				"esbenp.prettier-vscode",
-				"dbaeumer.vscode-eslint"
-			]
-		}
-	},
-	"overrideCommand": false
-}
