+++ ws/models/user.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+
+
+class UserCreate(BaseModel):
+    username: str
+    email: str
+    password: str
+
+
+class UserLogin(BaseModel):
+    username: str
+    password: str
